requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0
openai>=1.14.0
PyYAML>=6.0.1
python-telegram-bot[job-queue]>=21.0.1
pypdf>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5